Move all knowledge about tagging to the val module.
Although it's unlikely we'd want to change the scheme in a fundamental way, if
it's easier to understand if it's not spread across two modules.
