+++ src/lib/vm/val.rs
@@ -161,6 +161,38 @@ impl Val {
         }
     }
 
+    /// Create a (possibly boxed) `Val` representing the `isize` integer `i`.
+    pub fn from_isize(vm: &VM, i: isize) -> Result<Val, VMError> {
+        let top_bits = i as usize & (TAG_BITMASK << (BITSIZE - TAG_BITSIZE));
+        if top_bits == 0 || top_bits == TAG_BITMASK << (BITSIZE - TAG_BITSIZE) {
+            // top_bits == 0: A positive integer that fits in our tagging scheme
+            // top_bits all set to 1: A negative integer that fits in our tagging scheme
+            Ok(Val {
+                val: ((i as usize) << TAG_BITSIZE) | (ValKind::INT as usize),
+            })
+        } else {
+            Int::boxed_isize(vm, i)
+        }
+    }
+
+    /// Create a (possibly boxed) `Val` representing the `usize` integer `i`. Notice that this can
+    /// fail if `i` is too big (since we don't have BigNum support and ints are internally
+    /// represented as `isize`).
+    pub fn from_usize(vm: &VM, i: usize) -> Result<Val, VMError> {
+        if i & (TAG_BITMASK << (BITSIZE - TAG_BITSIZE)) == 0 {
+            // The top TAG_BITSIZE bits aren't set, so this fits within our pointer tagging scheme.
+            Ok(Val {
+                val: (i << TAG_BITSIZE) | (ValKind::INT as usize),
+            })
+        } else if i & (1 << (BITSIZE - 1)) == 0 {
+            // One of the top TAG_BITSIZE bits is set, but not the topmost bit itself, so we can
+            // box this as an isize.
+            Int::boxed_isize(vm, i as isize)
+        } else {
+            Err(VMError::CantRepresentAsIsize)
+        }
+    }
+
     /// If possible, return this `Val` as an `isize`.
     pub fn as_isize(&self, vm: &VM) -> Result<isize, VMError> {
         match self.valkind() {
@@ -234,6 +266,74 @@ mod tests {
         objects::{Class, ObjType, String_},
         vm::VM,
     };
+
+    use std::ops::Deref;
+
+    #[test]
+    fn test_isize() {
+        let vm = VM::new_no_bootstrap();
+
+        let v = Val::from_isize(&vm, 0).unwrap();
+        assert_eq!(v.valkind(), ValKind::INT);
+        assert_eq!(v.as_usize(&vm).unwrap(), 0);
+        assert_eq!(v.as_isize(&vm).unwrap(), 0);
+
+        let v = Val::from_isize(&vm, -1).unwrap();
+        assert_eq!(v.valkind(), ValKind::INT);
+        assert!(v.as_usize(&vm).is_err());
+        assert_eq!(v.as_isize(&vm).unwrap(), -1);
+
+        let v = Val::from_isize(&vm, isize::min_value()).unwrap();
+        assert_eq!(v.valkind(), ValKind::GCBOX);
+        assert_eq!(v.as_isize(&vm).unwrap(), isize::min_value());
+        let v = Val::from_isize(&vm, isize::max_value()).unwrap();
+        assert_eq!(v.valkind(), ValKind::GCBOX);
+        assert_eq!(v.as_isize(&vm).unwrap(), isize::max_value());
+
+        let v = Val::from_isize(&vm, 1 << (BITSIZE - 1 - TAG_BITSIZE) - 1).unwrap();
+        assert_eq!(v.valkind(), ValKind::INT);
+        assert_eq!(
+            v.as_usize(&vm).unwrap(),
+            1 << (BITSIZE - 1 - TAG_BITSIZE) - 1
+        );
+        assert_eq!(
+            v.as_isize(&vm).unwrap(),
+            1 << (BITSIZE - 1 - TAG_BITSIZE) - 1
+        );
+
+        let v = Val::from_isize(&vm, 1 << (BITSIZE - 2)).unwrap();
+        assert_eq!(v.valkind(), ValKind::GCBOX);
+        assert_eq!(v.as_usize(&vm).unwrap(), 1 << (BITSIZE - 2));
+        assert_eq!(v.as_isize(&vm).unwrap(), 1 << (BITSIZE - 2));
+    }
+
+    #[test]
+    fn test_usize() {
+        let vm = VM::new_no_bootstrap();
+
+        let v = Val::from_usize(&vm, 0).unwrap();
+        assert_eq!(v.valkind(), ValKind::INT);
+        assert_eq!(v.as_usize(&vm).unwrap(), 0);
+        assert_eq!(v.as_isize(&vm).unwrap(), 0);
+
+        let v = Val::from_usize(&vm, 1 << (BITSIZE - 1 - TAG_BITSIZE) - 1).unwrap();
+        assert_eq!(v.valkind(), ValKind::INT);
+        assert_eq!(
+            v.as_usize(&vm).unwrap(),
+            1 << (BITSIZE - 1 - TAG_BITSIZE) - 1
+        );
+        assert_eq!(
+            v.as_isize(&vm).unwrap(),
+            1 << (BITSIZE - 1 - TAG_BITSIZE) - 1
+        );
+
+        assert!(Val::from_usize(&vm, 1 << (BITSIZE - 1)).is_err());
+
+        let v = Val::from_usize(&vm, 1 << (BITSIZE - 2)).unwrap();
+        assert_eq!(v.valkind(), ValKind::GCBOX);
+        assert_eq!(v.as_usize(&vm).unwrap(), 1 << (BITSIZE - 2));
+        assert_eq!(v.as_isize(&vm).unwrap(), 1 << (BITSIZE - 2));
+    }
 
     #[test]
     fn test_recovery() {
